refactor: simplify tester implementation
